rename: reducer와 action 파일 분리
