+++ store/reducers/actionsType.js
@@ -1,0 +1,4 @@
+// classCalendar Type
+export const SET_MONTH = "SET_MONTH";
+export const SET_YEAR = "SET_YEAR";
+export const SET_IS_ADMIN = "SET_IS_ADMIN";
