+++ store/rootReducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
 import counterReducer from "./reducers/counterReducer";
 import authReducer from "./reducers/authReducer";
-import classCalendarReducer from "./reducers/classCalendarReducer";
+import classCalendarReducer from "./reducers/classCalendarReducer/classCalendarReducer";
 import modificationHistoryReducer from "./reducers/modificationHistoryReducer";
 import paymentHistoryReducer from "./reducers/paymentHistoryReducer";
